app/requirements.txt
python-decouple==3.8
python-dotenv==1.0.1
gunicorn==21.2.0
django==4.1.4
psycopg2==2.9.9
django-environ==0.11.2
whitenoise==6.8.2
psycopg2-binary==2.9.1